+++ CHANGELOG.md
@@ -6,18 +6,23 @@ Categories for each release: Added, Changed, Deprecated, Removed, Fixed, Securit
 
 ## Unreleased
 
-## [370.0] - beta
+## [370.1] - beta
 
 ### Added
 
 * Python 3.12 support
 * `--nextflow-pipeline-params` for `dx build --nextflow --cache-docker`
 * `dx build --nextflow --cache-docker` supports the `--profile` parameter to cache images associated with specific profile 
-* additional regional resources can be specified in dxworkflow.json or using `--extra-args` when `dx build` global workflows 
 
 ### Fixed
 
 * Released Nextaur 1.7.1. It contains multiple fixes in resolving docker image digests, handling file path collision, and OOM when creating new input stream.
+
+## [369.1] - 2024.2.13
+
+### Added
+
+* Additional regional resources can be specified in dxworkflow.json or using `--extra-args` when `dx build` global workflows 
 
 ## [369.0] - 2024.2.5